requirements.txt
# requirements file
# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision

#for example:
#torch==2.0.1
#matplotlib==3.7.2

crewai==0.75
crewai_tools==0.13.2
#yaml
PyYAML
dotenv
psycopg2
streamlit
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.5.0
plotly